Cleaned up CSS and some components
