+++ web/frontend/src/components/Footer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import "../styles/Footer.css";
+import "./Footer.css";
 
 const Footer = () => (
   <div className="flex flex-row justify-center items-center bg-gray-800 w-full p-4 text-gray-300 text-xs">
