+++ web/frontend/src/components/ProfilePicture.tsx
@@ -1,0 +1,37 @@
+import React from "react";
+
+const ProfilePicture = () => {
+    return (
+        <svg
+            className="w-7 p-1"
+            width="100%"
+            height="100%"
+            viewBox="0 0 741 809"
+            version="1.1"
+            style={{
+                fillRule: "evenodd",
+                clipRule: "evenodd",
+                strokeLinecap: "round",
+                strokeLinejoin: "round",
+                strokeMiterlimit: "1.5",
+            }}
+        >
+            <g transform="matrix(1,0,0,1,-1713.42,-904.592)">
+                <path
+                    d="M1753,1673.45C1753,1488.91 1901.26,1339.09 2083.88,1339.09C2266.49,1339.09 2414.75,1488.91 2414.75,1673.45"
+                    style={{ fill: "none", stroke: "white", strokeWidth: "79.17px" }}
+                />
+            </g>
+            <g transform="matrix(1,0,0,1,-2033.78,-719.63)">
+                <circle
+                    cx="2404.23"
+                    cy="913.551"
+                    r="154.338"
+                    style={{ fill: "none", stroke: "white", strokeWidth: "79.17px" }}
+                />
+            </g>
+        </svg>
+    );
+};
+
+export default ProfilePicture;
