+++ web/frontend/src/components/election/Ballot.css
