+++ web/frontend/src/components/election/BallotsTable.css
