+++ web/frontend/src/components/election/Election.css
