+++ web/frontend/src/components/election/ElectionDetails.css
