+++ web/frontend/src/components/election/Result.css
