+++ web/frontend/src/components/election/ResultPage.css
