+++ web/frontend/src/components/election/Status.css
