--- web/frontend/src/styles/ConfirmModal.css
@@ -1,42 +1,0 @@
-.modal-background{  
-    position: fixed;
-    top: 0;
-    left: 0;
-    width:100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.6);
-    z-index: 1;
-}
-
-.modal-wrapper{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    position:fixed;
-    background: #f6f6fa;
-    max-width: 100%;
-    max-height: 100%;
-    min-height:15%;
-    min-width: 10%;
-    top:50%;
-    left:50%;
-    transform: translate(-50%,-50%);
-    outline: #a9a9aa;
-    outline-style: solid;
-    outline-width: 1px;
-
-}
-
-.text-container{
-    padding: 5%;
-}
-
-.buttons-container{
-    overflow: hidden;
-    background-color: #177368; /* #177368 or #9bc0bc */
-    padding: 3%;
-}
-
-.btn-right{
-  float:right;
-}
